Optimise struct members for storage

+++ src/interfaces/IFundToken.sol
@@ -7,7 +7,7 @@ interface IFundToken is IERC20 {
     function mint(address recipient, uint256 amount) external;
     function burnFrom(address account, uint256 amount) external;
     function chargeManagementFee(uint8 mgtFee, uint256 price) external;
-    function distribute(uint256 distId, string memory distType, uint256 amount, uint256 scale) external;
+    function distribute(uint16 distId, string memory distType, uint256 amount, uint8 scale) external;
     function isCommitToken() external view returns (bool);
     function pause() external;
     function unpause() external;
